Create README.md
Hello. My name is Denise Kulawik.  I have an extensive and successful background in fundraising and development communications.  I'm new to technical writing and am especially interested in working on proposal development in the context of developing, funding, and selling technical products and services, especially in the healthcare and A&E sectors.   I can be reached at 323.509.2232.